using linear values

+++ src/main/java/com/zagorskidev/carpricing/domain/CarParams.java
@@ -5,24 +5,24 @@ import java.util.Map;
 
 public class CarParams
 {	
-	private Map<String, Integer> params;
+	private Map<String, Double> params;
 	
 	public CarParams()
 	{
 		params = new HashMap<>();
 	}
 	
-	public void setParam(String name, Integer value) 
+	public void setParam(String name, Double value) 
 	{
 		params.put(name, value);
 	}
 	
-	public void setParams(Map<String, Integer> params) 
+	public void setParams(Map<String, Double> params) 
 	{
 		this.params = params;
 	}
 	
-	public Map<String, Integer> getParams()
+	public Map<String, Double> getParams()
 	{
 		return params;
 	}
