+++ src/main/java/com/zagorskidev/carpricing/service/regression/Classifier.java
@@ -1,8 +1,6 @@
 package com.zagorskidev.carpricing.service.regression;
 
-import java.math.BigDecimal;
-
 public interface Classifier 
 {
-	public BigDecimal predict() throws Exception;
+	public Double predict() throws Exception;
 }
